properly support custom types
