+++ docs/data-sources/custom_name.md
@@ -1,26 +1,18 @@
 ---
 # generated by https://github.com/hashicorp/terraform-plugin-docs
-page_title: "namep_azure_name Data Source - terraform-provider-namep"
+page_title: "namep_custom_name Data Source - terraform-provider-namep"
 subcategory: ""
 description: |-
-  This data resource defines a name for an azure resource.
+  This data resource defines a name for a custom resource.
   The format will be used based on the the resource type selected and the appropriate format string.
 ---
 
-# namep_azure_name (Data Source)
+# namep_custom_name (Data Source)
 
-This data resource defines a name for an azure resource.
+This data resource defines a name for a custom resource.
 The format will be used based on the the resource type selected and the appropriate format string.
 
-## Example Usage
 
-```terraform
-data "namep_azure_name" "example" {
-  name     = "example"
-  location = "westeurope"
-  type     = "azurerm_resource_group"
-}
-```
 
 <!-- schema generated by tfplugindocs -->
 ## Schema
